spark-hub-algorithm/paper_parse/requirements.txt
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
nltk==3.8.1 